update to TS, remove tracked package-lock
